Create a separate custom schema for dashboard samples
